requirements.txt
speedtest-cli==2.1.3
fastapi==0.111.0
uvicorn==0.29.0
python-telegram-bot==21.3
win10toast==0.9
reportlab==4.2.0
pywin32==311
tenacity==8.3.0
schedule==1.2.2
requests==2.32.3
python-dotenv==1.0.1
